`ProcessLauncher`: except processes whose class cannot be imported
The `ProcessLauncher` would hit an exception if the class of the process
it is continuing could not be imported. The `plumpy.DefaultObjectLoader`
would throw a `ValueError` in this case which was not caught and would
simply except the future. The process task would be acknowledged and so
would not pose a problem, but the corresponding process node would not
be properly wrapped up and would keep the `CREATED` process state for
ever. The exception is now properly caught and the node is cleaned up
before being sealed.

Note that normally this is done by the process engine, but since the
exception happens before the process instance can be recreated this
pathway is never hit and we have to clean up the node manually. Since
the `ValueError` is not the only exception path possible, we catch a
broad exception and check if the node has already been excepted. If not,
the exception was raised before the process was recreated and we have to
perform the manual clean up.