Change `verdi computer list` to show unconfigured machines by default
Unconfigured machines used to be hidden by default, like disabled ones,
and could be displayed by using the `--all` flag. However, this was
shown to be unintuitive for new users who after creating a new machine
with `verdi computer setup` would not see it in the output of the
`verdi computer list` command.