Implement newly formalized graph traversal rules for export (#3262)
Thew newly formalized graph traversal rules are implemented in the
export interface. All twelve rules are used in traversing the graph
based on an initial set of nodes, but only a few can have the default
value overridden. Certain rules were decided not to be changeable and so
passing them will result in an exception. Also the command line interface
only exposed the toggleable flags.

The graph traversal logic has been moved from the export function itself
to a utility called `retrieve_linked_nodes`. Some duplication of queries
has been removed to improve the efficiency.

Archives now must be self-consistent in terms of the provenance graph
it represents, i.e., all nodes mentioned in the list of links have to
also be present in the archive. The danglink link checks can therefore
be removed from the export and import logic.