Fix loading of database environment in ReadTheDocs environment
The environment of ReadTheDocs does not actually have a database and so
the typical configuration and profile loading had to be hotwired. A
special global flag `IN_RTD_DOC_MODE` in `aiida.manage.configuration`
would circumvent loading an actual configuration for a dummy one. Since
the actual database environment never had to be loaded, the fact that
this contained non-sensical database connection parameters was
irrelevant. However, recently the documentation was changed to use the
Sphinx extension to document the `CalcJob` class. This now requires
loading the `aiida.orm` module, which in turn requires the database
environment being loaded.

The fix is to actually load the database environment through the backend
manager, but without performing the schema check. This will work despite
the unusable database configuration because no actual connection is
established in this way.