Properly use nested transaction when deleting nodes in SqlAlchemy
The `delete_nodes_and_connections_sqla` function was opening a new
session and committing the changes, leaving other sessions, for example
the unit test one, desynchronized. This would cause `DetachedInstance`
exceptions to be thrown. The solution is to get a properly nested
transaction which we get through the `Backend.transaction` context
manager.