Enable auto-completion for `IdentifierParamType` (#3375)
The `OrmEntityLoader` class is extended to allow specifying a query
operator in the `get_query_builder` method. This allows the introduction
of the `get_options` method, which unlike `load_entity` will not try to
load one specific entity, but provide a list of possibilities for a
given identifier. This in turn allows the CLI `IdentifierParamType` to
enable auto-completion. Here we only enable it for `LABEL` identifiers
and do not try to infer the identifier type. Finally, the new
functionality is hooked up for the `ComputerParamType`.