Update delete node functionality to respect formalized graph traversal rules (#3367)
With the graph traversal rules for node deletion now well defined, the
`delete_nodes` utility is updated accordingly. Certain link traversal
flags had to be changed in the utility function as well as the `verdi`
command `verdi node delete`.

The new unit tests uncovered a problem with the backend utility function
`delete_nodes_and_connection` for the SqlAlchemy backend. This was itself
opening a new session to delete the nodes and links. Since the session
is not nested within the current test session, the deletions were not
effectuated in the latter resulting in `DetachedInstanceError`. By using
the `transaction` context manager of the backend instance, the session
will be properly nested.