Fix randomly failing REST API test (#3292)
The test was assuming that the results were ordered by pk as the data
with which it is compared is ordered as such. However, the query did
not specify this ordering explicitly, unlike all other tests. This means
that the ordering was random. Often this happens to be ordered by pk
causing the test to pass, but sometimes by chance it was not causing the
random failures.