Prune input mapping of `ProcessBuilder` before creating `Process`
The validation of a `ProcessSpec` will trigger for a port as soon as a
value has been passed. For a port namespace this is also the case, even
if it contains no actual values. This was a problem when launching a
process from a `ProcessBuilder`. If it contained an optional namespace,
it would always pass at least an empty dictionary for said port, even if
the user did not specify any specific input within it. This would then
always trigger the validation which would fail. This limitation made the
concept of optional port namespaces fundamentally incompatible with the
`ProcessBuilder`. Here we add a method to "prune" the input mapping of a
builder before it is being passed to the process constructor. This will
remove any portnamespace that is completely empty, meaning it contains
no other values other than other empty dictionaries (namespaces).