Add test on Travis to check the loading time of `verdi`
The loading time of `verdi` needs to be kept to a minimum in order to
keep it snappy and tab-completion from workable. This means that the
database environment should not be loaded until it is absolutely
necessary and only within the function body of commands. This means that
`aiida.orm` cannot be imported in the top level of any `aiida.cmdline`
file, as that will_ trigger the loading of the database environment.
However, since this is not easy to spot, this happens all the time,
breaking `verdi` essentially.

We add a test to the Travis build that will use `time` to measure the
loading time. If the given limit is exceeded a certain number of times
consecutively, the test will fail. The limit is set rather loosely,
because we do not want numerical noise to fail this test. The typical
mistake causing the loading time to increase, should largely exceed the
limit so should always be caught.