Centralize definition of graph traversal rule toggleability and default (#3394)
The available graph traversal rules are identical for node export and
deletion. Only the toggleability and defaults differs for each of the
two scenarios. To prevent mistakes, we centralize their definition in
one location. Two dictionaries in `aiida.common.links` provide a mapping
of each traversal rule onto a named tuple that records the default value
and toggleability.

This notation allows for dynamic code, for example, the command line
options that toggle the rules are now generated dynamically for the
`verdi export create` and `verdi node delete` commands.

In addition, the old `LINK_FLAG` data structure was defined in the
`aiida.tools.importexport` module which indirectly causeed `verdi` to
slow down. The dictionary was needed to generate the command options
and by importing it, indirectly the `aiida.orm` module was imported.