Ensure unpacked archive directories are not deleted on import
Since the `extract_tree` function did not actually copy the contents of
the source directory in the operational `SandBoxFolder` the context
manager of the `Archive` class would delete the source directory itself
after import. By simply copying the contents of the source into the
sandbox before starting the import, the source is left untouched. This
also allows to have the same code pathways for all archive formats,
independent of whether it is a packed file, like zip, or not.

Co-Authored-By: Casper Welzel Andersen <casper.andersen@epfl.ch>