Move and improve `CalcJob` input validation to process spec
Compound input validation for the `CalcJob` was mostly implemented on
the node level in `CalcJobNode._validate`. This has been moved to a
validator on the input portnamespace of the `CalcJob` in order to catch
inputs problems as soon as possible. It implements existing validation:

 * A `Computer` has been defined
 * The parser class, if specified, can be successfully loaded
 * The specified resources are valid and sufficient for the scheduler

In addition, the validation on the `Computer` has been improved. It now
checks that if it is specified both directly in the `metadata` as well
as indirectly through the `Code` input (in the case of a remote code)
that it concerns the same computer.