Add `export_parameters` to export archive `metadata.json` (#3386)
The new dictionary contains the parameters used in creating the export
file that can be useful at later stages to debug in the case of archives
that seem inconsistent. Especially the `graph_traversal_rules` that
detail how the originally selected nodes of the `entity_starting_set`
were extended through the graph traversal algorithm.