Ensure `Data` export methods return correct type
Sub classes of `Data` can implement methods that export their content in
a certain format. The `Data.export` function will call through to these
sub class specific methods and expects the returned content to be a
byte string. This was not always the case and so caused exceptions. A
test is added that checks the type of the returned tuple for the `Data`
sub classes that currently have export functionality.

The `verdi data * show/export` commands that serve as a CLI endpoint for
this export functionality will have to make sure that if they need to
write the returned export content to file, that they open the target in
binary mode. This was not the case for all formats and is now fixed.