Use correct session to reset database in SqlAlchemy unit tests
The method that reset the contents of the database after each set of
unit tests for the SqlAlchemy backend was retrieving a session that
would be disconnected from the current test session. This would cause
then leave orphaned ORM instances in the main session that no longer
corresponded to actual database instances resulting in exceptions like
`ObjectDeleteError`. Going through the backend to retrieve a transaction
will ensure that the session is nested in the main test session.