Disallow pickle when storing numpy array in ArrayData
Unpickling untrusted data is a potential security risk, since it
allows arbitrary code execution. For this reason, numpy changed
its 'load' method to no longer allow pickles by default. This
means that in AiiDA, ArrayData containing pickles can be stored,
but not loaded. To make things consistent, we now also disallow
pickle when saving the array. Besides the security risk, pickle
is also problematic in the context of AiiDA because it can also
become unreadable e.g. when the classes it contains change names.