Make spacing consistent by removing <p> from single description.
So far, the description for a single port was wrapped inside both
a paragraph and a list item. This caused inconsistent vertical
spacing when namespaces and single ports were mixed. To avoid this,
the paragraph wrapping the content was removed. If the increased
spacing is desired, this should probably be done in CSS on the
list.