Collapsible namespace in sphinxext (#3441)
* Create a 'details' HTML tag for the content of a Namespace.

Note that this only works in Python3, because the extension used
to create the <details> tag is only compatible for Sphinx>=2.0.

* Make spacing consistent by removing <p> from single description.

So far, the description for a single port was wrapped inside both
a paragraph and a list item. This caused inconsistent vertical
spacing when namespaces and single ports were mixed. To avoid this,
the paragraph wrapping the content was removed. If the increased
spacing is desired, this should probably be done in CSS on the
list.

* Update sphinx dependency to 2.2.0 for python3

Because the sphinxcontrib-details-directive requires sphinx>=2.0,
this dependency is also updated. Since that version of Sphinx is
incompatible with Python2, a version selector is added, and for
Python2 the previous version of Sphinx is kept.

* Change doc highlighting to 'ipython', update sphinxcontrib-contentui

The 'python' highlighting does not work in the visualising_graphs
page, because an ipython ! command (shell command) is used. Changing
the highlighting type to 'ipython' fixes this issue.
The sphinxcontrib-contentui needs updating to be compatible with Sphinx
version 2.2.0.