Include more detail in process state of node summary (#3444)
The node summary is used by `verdi node show` and `verdi process show`
to display basic information about the node. The process state was
already included, but is often not enough. If the process is excepted we
add the exception message and in case of a failed process the exit
message associated with the exit code.