Define custom object loader that raises `ImportError` (#3442)
The default object loader from `plumpy` was raising `ValueError`
exceptions if the process class could not be loaded. However, catching
the `ValueError` in the `ProcessLauncher._continue` is a bit broad if we
only want to target problems with loading the process class. Therefore
we create our own `ObjectLoader` that raises a `ImportError` instead.