Fix the `is_process_function` attribute of process functions (#3454)
The `is_process_function` attribute was set to a property that would
always return `True`, except the way it was implemented calling it on a
decorated function would just return the property object and not the
value it should return. This bug went unnoticed even in the utility
`aiida.engine.utils.is_process_function` which would return the property
of a wrapped function, which since that is considered "truthy" would
still have the desired effect, but for the wrong reasons. There is no
need to have this be a property, so instead we simply set it as an
attribute. The utility will now return a value that also explicitly
checks for the `True` boolean value.