Add generator of full available node namespace
A utility `aiida.restapi.common.identifiers.get_node_namespace` is added
that constructs the full nested node namespace present in the current
database. Each entry will contain a `full_type` attribute, which
uniquely identify the exact type, in the case of a leaf, or the sub
class of node types in the case of a namespace. The `full_type` can be
used to generate the query builder filters that will allow to query
exactly for this node type or its sub classes. This is stored in the
REST API module because it is exposed as an endpoint. This can then be
consumed by the Materials Cloud to dynamically build up an index of
available node classes with their `full_types` which can be used to
query for them.