REST API v4: updates to conform with `aiida-core==1.0.0`
The REST API is updated to `v4` to reflect the changes of the ORM of
`aiida-core` in `v1.0.0`, especially in the available endpoints. The
`calculations` and individual `data` endpoints have been removed. Added
are the `processes` and `calcjobs`  endpoints. The `visualization`
endpoints, which were used by the Materials Cloud to retrieve data in
a specific format that was derived from the basic node attributes and
repository contents, has been changed to `derived_properties`.

Also added the endpoints `repo_list` and `repo_contents` to get the
available objects in a node's repository and their contents in the case
of files.

The `schema` endpoints have been replaced with `projectable_properties`.
The `get_schema` class methods on various ORM classes have been
deprecated as this functionality was used strictly by the REST API.

Co-Authored-By: epassaro <elsa.passaro@epfl.ch>