remove autouse=True from clear_database fixture
With `autouse=True`, there is no straightforward way for plugin authors
to disable use of a fixture.
While this makes sense for the `aiida_profile` fixture, we've
encountered cases where resetting the database in between tests may lead
to issues.

This problem is solved by keeping the fixture available but requiring
users to depend on it explicitly.