Add type checks to all plugin factories (#3456)
The various entry point factories, based on the base class
`aiida.plugins.factories.BaseFactory` would simply return the resource
that the given entry point corresponds to if it could be loaded. There
were no additional checks if the type of the loaded entry point
corresponds to what is expected for that particular entry point group.
It was possible, for example, to add `CalcJob` implementations and
calculation functions to the `aiida.workflows` entry point group, as well
as `WorkChain` and work functions to `aiida.calculations`. The respective
factories would load these entry points without complaining. However, to
keep consistency it is better to raise if an entry point is loaded from
the incorrect entry point group.

The functionality to check that all registered plugins can be loaded
through their respective factories has been abstracted to a utility in
`aiida.plugins.entry_point.validate_registered_entry_points`. It is also
exposed on the command line through `verdi devel validate-plugins`.