Emit a warning when input port specifies a node instance as default
Using mutable objects as defaults for `InputPorts` can lead to
unexpected result and should be discouraged. Implementing this for all
types, including python builtins such as lists and dictionaries, is
difficult and not the most pressing problem. The biggest problem is
users specifying node instances as defaults. This will cause the backend
to be loaded just when the process class is imported, which can cause
problems during the building of documentation and with unit testing
where instances in memory to deleted nodes can be referenced, among
other things.

To warn users for these complications, we override the `InputPort`
constructor to check the type of the default and emit a warning if it
is a node instance.

The `CalcJob` implementation had to be adapted to conform with the new
requirements as the `mpirun_extra_params` and `environment_variables`
metadata options were using mutable types for their default. Just as
certain test process classes, the defaults are now prefixed with the
`lambda` keyword.