Refactor unit test managers and add basic fixtures for `pytest` (#3319)
The utilities for plugin testing are reorganized from a single file
`aiida.manage.fixtures.py` over multiple files. The central manager class
is now `TestManager` in the `aiida.manage.tests` package. Test framework
specific code is moved to `unittest_classes.py` and `pytest_fixtures.py`
respectively, the latter of which contains a few basic fixtures that can
be easily used in combination with `pytest.

The `TestManager` can now use either a `TemporaryProfileManager` or an
instance of `ProfileManager` which will use an existing test profile.
This gives the flexibility to choose the ease of the automatic profile
generation with entire database stack, or if one prefers performance an
existing profile can be used. For use with `pytest` the desired test
profile can be set by defining the `AIIDA_TEST_PROFILE` environment
variable.