Add database data migration to seal unsealed processes (#3470)
This should have been taken care of by the previous migration named
`0040_data_migration_legacy_process_attributes`, however, the query was
incorrect and would skip processes that did not have a `sealed` attribute
nor a `process_state` attribute, which is the case for certain legacy
calculations like `InlineCalculation`. The fix is that in the `where`
clause first check for the presence of the `process_state` attribute
before performing the check that its value does not correspond to an
active state.