Backport database schema generation check
This check will be implemented in the code after the schema has been
reset, but to make sure that people with a database schema generation
that is newer than the current one, cannot trigger the database, we
backport the check here. Note that this will only affect the release of
`v1.0.0 which will include this commit. Older versions will just attempt
the migration, which will fail and except. So at least it should not
touch and corrupt the database.