Fix bug in `upload_calculation` when running with `store_provenance=False`
When running a `CalcJob` in dry-run mode, it is often advised to also set
`store_provenance=False` which means nodes will not be stored, so input
nodes can be unstored. However, in `upload_calculation` to process the
`local_copy_list` the `load_node` function was used to load the
corresponding node, which would fail in this scenario. The only possible
solution is to pass the full mapping of inputs to the function and if
the loading of the node fails, try to fish out the corresponding node
from the mapping. If all fails, we emit a warning and skip the rule.