`QueryBuilder`: add deepcopy implementation and `queryhelp` property (#3524)
A new `QueryBuilder` instance can be constructed in full from the query
help. This can be retrieved through `get_json_compatible_queryhelp` for
which we add a `queryhelp` property as shortcut`. This is in turn used
for the implementation of `__deepcopy__` to allow to clone a builder
instance. The old tests in `test_query.py` have been merged with the ones
in `orm.querybuilder`.