Enable CI on Github
Add a GitHub workflow `.github/workflows/ci.yml` that will trigger on
push and pull requests. It will run the five actions it defines:

 * `conda`: will install conda and create the environment
 * `docs`: builds the documentation with nitpick warning
 * `pre-commit`: runs pre-commit on all files
 * `tests`: runs `verdi devel tests` and the stand alone test files
 * `verdi`: runs the tests that check the load time is acceptable

All tests are performed for python 3.7 except for the `tests` action
that is done for 3.5 as well. Both python versions are run for both
database backends in a matrix strategy. Even though we support 3.6 as
well we do not explicitly test it since it will require another two
builds and testing 3.5 and 3.7 should give decent guarantees.

Finally the argument for multiple individual actions instead of joining
them is based on the fact that there does not seem to be a limit on
concurrent number of actions on GitHub as of this writing. This means
that by spreading them out, allows running them in parallel which should
reduce the overal runtime of the continuous integration workflow.