Add data migration to rename legacy default return link label
The original default link label that was used when no explicit one was
specified by the user was set to `_return`. In a later change, link
labels starting or ending with an underscore were banned because it
would interfere with use of double underscores to indicate namespaces.
As a result the default link was renamed to `result` however a data
migration was never put in place.