Fix REST API `attributes_filter` and `extras_filter` (#3556)
The REST API `attributes_filter` and `extras_filter` were not working
correctly when pagination is requested. Without pagination they return
attributes and extras dictionaries with the keys specified in
`attributes_filter` and `extras_filter` as expected. When pagination is
added they return instead the attributes keys as `attributes.key`, which
is not consistent with the expected behaviour.

Changed REST API version to 4.0.1.