Create backup of configuration file before migrating
This allows users to revert in case of problems with the migration or if
the migration happens by accident. Since the migration is done
automatically without prompt and is not reversible, having a backup is
crucial if users checkout a version of the code by accident.

To accomplish this the configuration instantiation and migration is also
refactored moving the responsibility of migrating a configuration file to
the `Config` class itself. Constructing a `Config` instance from an
existing configuration file should now be done through the `from_file`
class method. This will backup the file if it determines that it is
outdatedand then construct a `Config` instance from the migrated contents.