Remove Travis configuration and simplify Jenkins setup
The majority of tests are now run through a GitHub CI workflow. This
means it is not necessary to also run those on Jenkins, especially that
Jenkins was originally configured to just run additional tests that were
more computationally intensive. Therefore, we only keep the RPN test to
run on Jenkins for now.