Fix problem with dummy configuration file on ReadTheDocs (#3579)
The recent change in the backup behavior of the configuration file leads
to problems on ReadTheDocs. The dummy config returned by `load_config`
when on RTD uses `/dev/null` as dummy filepath, but the new backup
mechanism will try to write a backup using that filepath as a prefix,
leading to a filepath that can't be written to. The temporary fix is to
use an actual temporary file for the dummy filepath.