Further improvements to `verdi computer test` output
The number of tests is increased and messaging from within each test is
disabled. The tests now return a tuple of a boolean indicating success
or failure and an optional message. This message can be informational in
case of success or an error message. With echo'ing only being performed
in the main function, the output can be given a consistent look. Each
test will have `[OK]` or `[FAILED]` on the same line with an optional
message. Additional lines, such as for the full traceback, are
prepended with newline and two spaces of indentation.