print path to config file in verdi status
From time to time, the need arises to have a look inside the currently
active AiiDA configuration file.
While the path to the file repository is printed in the output of `verdi
status` and `verdi profile show`, one needs to edit this path to get to
the config file (and, in principle, the repository can even be located
comepletely elsewhere).

Printing the location of the config file is helpful for daily use, also
as a reminder which configuration file is currently active.