Drop Python 2: remove arguments from `super` calls
This is no longer necesary in python 3. To remove all the occurrences
of this usage, the following command was used:

    find . -type f -not -path './.git*' -exec \
    sed -i 's/super([^)]*,[^)]*)/super()/g' {} +

The regex tries to find all instances of `super()` where there is
content between the parentheses. The search for `[^)]*` means to look
for all characters except a closing parens, which essentially makes the
search non-greedy. Having just that would not be enough and we have to
explicitly add another such clause separated by a comma. This is the
exact format required by python 2 where the `super` call expects two
arguments, first being the class itself and the second the reference,
typically `self` or `cls. By making the regex more specific we avoid
matching cases like:

    def test_without_super(self):

which would be replaced to

    def test_without_super():

if we don't include the explicit comma in the regex between the parens.