Fix false positive for `verdi database integrity detect-invalid-links` (#3591)
The query `SELECT_CALCULATIONS_WITH_OUTGOING_CALL` defined in the module
`aiida.manage.database.integrity.sql.links` contained an error causing
the command to return false positives. Any node with an outgoing link
of type `CALL_WORK` was being flagged. The problem was not detected by
the tests because they operated on an empty database and only checked
for true positives. A basic graph is added to now detect the false ones.