fix print of exception in gotocomputer
"verdi calcjob gotocomputer" obviously does not work when no authinfo is
present for the computer to go to.
There was a bug in printing the exception, resulting in an exception
bubbling up to user space (instead of a critical error with a helpful
message).

The error message now looks like this:

$ verdi -p quicksetup calcjob gotocomputer 712
Critical: NotExistent('User<daniele.ongari@epfl.ch> has no configuration for Computer<fidis (Imported #0)>')