Relax requirements of required and optional dependencies
The version requirements of all `install_requires` and `extras_require`
dependencies are reevaluated. Where we used to require explicit version
by pinning to exact versions using the `==` operator, we now use the more
liberal operator `~=`. This is the compatible release operator:

    https://www.python.org/dev/peps/pep-0440/#compatible-release

For a given release identifier `V.N` the compatible release clause is
approximately equivalent to the pair of comparison clauses:

    >= V.N, == V.*

With this rule we define our dependency requirements using the minor
version for packages with a major version of 1 or higher and using the
patch version for packages whose major version is still at 0. The reason
is that when packages have released `v1` they can be expected to respect
semantic versioning and not break backwards compatibility in minor
version whereas this is typically not the case for `v0` versions where
backward incompatible changes are introduced in minor versions as well.
Note that this approach will not prevent from breaking our builds
because packages break their API in minor versions, but at least we have
a consistent dependency requirement policy.

List of specific restrictions and changes:

 * `pymatgen==2019.7.2` is the last to support python 3.5
 * `jinja2` moved to `install_requires` as it is used explicitly by `verdi`
 * `astroid==2.2.5` required by `prospector`, will cause crash if wrong

Explicit lower patch requirements are set for a few packages because
they are known to contain critical and required bug fixes:

 * `psycopg-binary>=2.8.3`
 * `sqlalchemy>=1.3.10`
 * `pgtest>=1.3.1`
 * `seekpath>=1.9.3`

Finally, an exception is made for packages of the `dev_precommit` extra
which can affect pre-commit behavior such as linting warnings or code
formatting. To prevent updates in these packages from unexpectedly
failing the pre-commit hooks, they are pinned to an explicit version.