Remove unnecessary explicit dependency `mock`
This is a rolling backport of the module `unittest.mock` which has been
part of the standard python library since v3.3.

Note that one has to do `from unittest.mock import patch` and use it
directly as for some reason `import unittest` followed by using
`unittest.mock.patch` will result in an `AttributeError`.