do not create config file upon import
So far, importing aiida would directly create a config file.
Such side effects can have problematic consequences e.g. if the default
path that AiiDA chooses is not writeable. It also means that you pollute
your file system just by importing the package.
This PR moves the creation of the config file (if necessary) into the
`load_profile` function, which is anyhow called e.g. by the `verdi`
command line.

The same goes for the configuration of the "logger".