Temporarily pin the requirement for `numpy`
The `numpy==1.18.0rc1` is broken and apparently does get installed with
the compatible version operator. Note that since it will be installed by
`pymatgen` during its setup phase, the version limitations in our
`install_requires` are ignored. We therefore have to make sure we
explicitly install the compatible `1.17.4.` before installing the rest
of the dependencies.