Remove unnecessary `aiida.common.utils.get_repository_path`
The reason for this utility function existing, was to cache the calls to
`os.path.abspath` which can be expensive. However, the function they
call to get the basepath `Profile.repository_path` already makes sure
that it returns an absolute path and is called only once. Since the
function call can be replaced with a single line and the caching is no
longer necessary and it wasn't tested, it is best to remove it.