Reduce the number of public methods of `Computer` front-end class
The `aiida.orm.computers.Computer` front-end class had a mixture of
properties and methods to get and set various properties. Which one was
available depended on how that property was stored in the backend. Those
that are stored as database columns were exposed through property getter
and setters, whereas properties that are part of the `metadata` JSONB
column had individual methods.

Besides being inconsistent, having explicit methods for each property,
of which there are already quite a few and which might increase in the
future, is inconvenient. It also makes it more difficult to
programmatically get all the properties that "define" a computer. By
defining all properties in a mapping class attribute, one can define all
setters and getters dynamically.