Add `--format` option to `verdi computer show`
This allows to, in addition to the default table display, dump the
contents of a computer in yaml format which can in turn be used for the
`--config` option of `verdi computer setup`. This makes the sharing of a
particular computer that has been setup very easy. Sharing yaml config
files is more apt than AiiDA export archives which was the only possible
way of sharing `Computer` instances to date.